+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 2026-3-9
+
+#### New:
+- 加入[Algorithm::ChthollyTree](Algorithm/ChthollyTree.h), 快速维护区间上的性质
+
+-----
 ### 2026-2-26
 
 #### New: